Design and implement JSONTree
Based on the [RFC 9535] JSONPath standard, create the jsontree package.
It supports all the features of the standard except for Filter
selectors, but organizes them as a tree structure rather than a linked
list. This allows queries to be designed that select a subset of a
structured JSON value while preserving the structure of the original
value.

Organized using one or more name (object key), index, wildcard, and
slice selectors in a segment. Execution collects all values for a given
segment in relation to the hierarchy of the structured JSON value and
replicates them in a destination value. All keys and array indexes are
preserved from the root of the structured value.

Much more to be done, including:

*   More documentation, especially on the behavior of array index
    selection
*   JSONPath merging, so one can define a query as a series of JSONPath
    queries and JSONTree will merge them into a single tree query.
*   Parsing of JSONPath expressions?

  [RFC 9535]: https://datatracker.ietf.org/doc/rfc9535/
    "JSONPath: Query Expressions for JSON"