Add support for filters
It's pretty straightforward using the github.com/theory/jsonpath `spec`
package, but since filters can operate on the root node as well as the
current node, update all of the execution methods in tree.go to pass
both the root node and the current node (renamed from `src` to `cur`),
then pass both to `spec.FilterSelector.Eval`.

Add tests to ensure filters stringify properly, and that filter
execution operates properly, both in tree structures and evaluating the
current and root nodes.

While at it, update the README to reflect the adoption of
github.com/theory/jsonpath to handle selector execution.