Implement `Segment.Contains`
This method returns true if a segment contains a selector. It supports
direct comparison of Name, Index, and stringified Filter selectors, but
also makes an effort to return true when a Slice selector contains an
Index or is a subset of another Slice.

Also: Add a stability notice to the README.