Implement path merging
Change `New` to accept a list of jsonpath.Path objects and teach it to
merge them into a single tree structure, using the Segment struct
defined here and relying on jsonpath.Selector for selectors. In order to
create as efficient a query tree as possible, examine each segment and
selector in all the paths and try to combine them, eliminate duplicates,
and generally simplify things wherever possible.

General notes on how `New` works:

Make an effort to minimize the number of selectors in each segment of
each path. Whereas `jsonpath.Select()` returns values for each selector
that matches, even dupes, `jsontree.Select` does not, since it preserves
the original data structure along the selected paths. So, for example,
wildcards eliminate the need for any other selectors, indexes and slices
that are encapsulated by other slices can be removed, and duplicates are
removed, including filters (matching on stringification).

The comparisons are imperfect. For example, two filters can be logically
equivalent but have different orders of operands, so would not be
considered equivalent. This may be rectified in the future by
normalizing filter stringification. Slice comparison is also sometimes
impossible, mainly when using negative index arguments, since the values
then become dependent on the length of input. These redundancies should
be acceptable, however, as relatively less common and just mean some
values would be selected twice. Their places in the resulting data
structure are unchanged, however, so the outcome will be the same.

Trim trailing wildcard segments, since the result is the same as
selecting the entirety of the parent segment.

Merge segments from different paths wherever their subsequent segments
are identical, or when their selectors are equivalent. The exception is
when one segment is a descendant segment (`..[]`) and the other is not.
In that case, discard the non-descendant segment only when both are
wildcards and constitute the same subpath.

Expand the selector and sub-path comparison logic in `segment.go`,
including comparison and merging of selectors (eliminating duplicates
and slice overlaps) and exact comparisons (ignoring slice overlaps).
Switch from methods to functions to reduce duplication between
equivalence vs. exact comparison and fill in a missing filter
comparison.

Add `isBranch`, which compares a list of `jsonpath.Segments` to a
`jsontree.Segment` and returns true when their paths are logically
equivalent. This means that the `jsontree.Segment` must have only one
sub-path with no additional branches.

Add `deduplicate` and run it after the paths have been merged, to make a
second pass over the entire tree structure and eliminate further
duplicates, notably segments with identical children, taking descendant
vs regular child selectors into account, and then merging slice
selectors.

Add tests for all of this functionality, covering all the paths in both
`TestNew` and unit tests for all the functions.

Finally:

*   Update the example test to use jsonpath.Path objects.
*   Rename the `JSONTree` struct to `Tree`