Implement ordered mode
This mode preserves the order, but not position of array values. It
contrasts with the existing mode, now named "fixed mode", which
preserves the index positions of array values. Add `NewFixedModeTree` to
create a Tree in fixed mode, and change `New` to use ordered mode.

It works by factoring array index setting into a new method,
`Tree.insert`, that replaces `nil` values with a `null` placeholder in
ordered mode. Then, `Select` recursively compresses all selected arrays
by removing `nil` values and replacing `null` placeholders with `nil`s.
This ensures that `null` values selected by paths remain in the result.

It would be nice to find a way to enable this behavior without iterating
over the entire selected value before returning it. An attempt to use a
struct that automatically handled both append and index-preserving modes
failed because arrays can be processed multiple times, but should not be
appended multiple times. Thus the solution here, which attempts to
minimize resource allocation at the expense of extra processing for
every `Select`.

Add test expectations for ordered mode and test them alongside fixed
mode. New tests also ensure that selected `nil`s are preserved in
ordered mode while unselected `nil`s are not.

Document the two modes and how they differ, noting that ordered mode is
the default. Add examples for each, and update the use cases
documentation to reflect the application of each mode.

While at it:

*   Re-use the underlying array when deriving a new array from its
    values, and use `slices.Clip` to remove unused capacity from the new
    slice.
*   Take advantage of this change to simplify the logic of
    removeCommonSelectorsFrom`.