Refactor segment stringification
`segment.String()` was not including the selectors for the current
segment in its output, only its children. This made debugging tricky.

So refactor segment stringification by adding `segment.writeSelectors`
and teaching `segment.String` and `segment.WriteTo` to use it. This
allows stringification of a selector to always include the selectors and
children.

In other words, this code:

``` go
child(spec.Wildcard).Append(
    child(spec.Index(0), spec.Index(1))
).String()
```

Previously output:

```tree
[0, 1]
```

And now outputs this complete representation of the segment:

```
*
└── [0, 1]
```

This change requires that `Tree.String` now also iterate over children,
so make that change, as well, and add explicit tests for `Tree.String`.