Improve slice merging
It was not previously exhaustive, because there was more than one
perviously-found super-slice, only one would be removed. Modify
`mergeSlices` to keep iterating over selected slices until none are
replaced. Discovered while adding playground links to the README, where
an example did not work as expected. Add that example as a test.